HSEARCH-4287 Move some tests out of DatabasePollingAutomaticIndexingEdgeCasesIT and into BackendFailureIT
Because we're going to add more "edge cases" and they will make the
model more complex.
