+++ integrationtest/mapper/orm/src/test/java/org/hibernate/search/integrationtest/mapper/orm/automaticindexing/coordination/databasepolling/DatabasePollingAutomaticIndexingEdgeCasesIT.java
@@ -6,41 +6,22 @@
  */
 package org.hibernate.search.integrationtest.mapper.orm.automaticindexing.coordination.databasepolling;
 
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.awaitility.Awaitility.await;
-import static org.hibernate.search.util.impl.integrationtest.mapper.orm.OrmUtils.withinTransaction;
-
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.TimeUnit;
 import javax.persistence.Basic;
 import javax.persistence.Entity;
 import javax.persistence.Id;
 
 import org.hibernate.SessionFactory;
-import org.hibernate.search.engine.reporting.EntityIndexingFailureContext;
-import org.hibernate.search.engine.reporting.FailureContext;
-import org.hibernate.search.engine.reporting.FailureHandler;
-import org.hibernate.search.engine.reporting.impl.LogFailureHandler;
 import org.hibernate.search.mapper.orm.coordination.CoordinationStrategyNames;
-import org.hibernate.search.mapper.orm.common.EntityReference;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.GenericField;
 import org.hibernate.search.mapper.pojo.mapping.definition.annotation.Indexed;
-import org.hibernate.search.util.common.SearchException;
 import org.hibernate.search.util.impl.integrationtest.common.rule.BackendMock;
 import org.hibernate.search.util.impl.integrationtest.mapper.orm.CoordinationStrategyExpectations;
 import org.hibernate.search.util.impl.integrationtest.mapper.orm.OrmSetupHelper;
 import org.hibernate.search.util.impl.integrationtest.mapper.orm.OrmUtils;
-import org.hibernate.search.util.impl.test.annotation.TestForIssue;
 
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
-
-import org.awaitility.core.ThrowingRunnable;
 
 /**
  * Extensive tests with edge cases for automatic indexing with {@link CoordinationStrategyNames#DATABASE_POLLING}.
@@ -59,14 +40,11 @@ public class DatabasePollingAutomaticIndexingEdgeCasesIT {
 			.enableFilter( false );
 
 	private SessionFactory sessionFactory;
-	private TestFailureHandler failureHandler;
 
 	@Before
 	public void setup() {
 		backendMock.expectSchema( IndexedEntity.INDEX, b -> b.field( "indexedField", String.class ) );
-		failureHandler = new TestFailureHandler();
 		sessionFactory = ormSetupHelper.start()
-				.withProperty( "hibernate.search.background_failure_handler", failureHandler )
 				.withProperty( "hibernate.search.coordination.processors.indexing.outbox_event_finder.provider", outboxEventFinder.provider() )
 				.setup( IndexedEntity.class );
 		backendMock.verifyExpectationsMet();
@@ -116,9 +94,6 @@ public class DatabasePollingAutomaticIndexingEdgeCasesIT {
 					);
 		} );
 		backendMock.verifyExpectationsMet();
-
-		assertThat( failureHandler.genericFailures ).isEmpty();
-		assertThat( failureHandler.entityFailures ).isEmpty();
 	}
 
 	@Test
@@ -146,265 +121,6 @@ public class DatabasePollingAutomaticIndexingEdgeCasesIT {
 			} );
 			backendMock.verifyExpectationsMet();
 		}
-
-		assertThat( failureHandler.genericFailures ).isEmpty();
-		assertThat( failureHandler.entityFailures ).isEmpty();
-	}
-
-	@Test
-	public void backendFailure() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-			entity1.setIndexedField( "initialValue" );
-			session.persist( entity1 );
-
-			IndexedEntity entity2 = new IndexedEntity();
-			entity2.setId( 2 );
-			entity2.setIndexedField( "initialValue" );
-			session.persist( entity2 );
-
-			IndexedEntity entity3 = new IndexedEntity();
-			entity3.setId( 3 );
-			entity3.setIndexedField( "initialValue" );
-			session.persist( entity3 );
-
-			CompletableFuture<?> failingFuture = new CompletableFuture<>();
-			failingFuture.completeExceptionally( new SimulatedFailure( "Indexing work #2 failed!" ) );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.createAndExecuteFollowingWorks()
-					.add( "1", b -> b
-							.field( "indexedField", "initialValue" )
-					)
-					.add( "3", b -> b
-							.field( "indexedField", "initialValue" )
-					)
-					.createAndExecuteFollowingWorks( failingFuture )
-					.add( "2", b -> b
-							.field( "indexedField", "initialValue" )
-					)
-					// retry (succeeds):
-					.createAndExecuteFollowingWorks()
-					.addOrUpdate( "2", b -> b
-							.field( "indexedField", "initialValue" )
-					);
-
-		} );
-		backendMock.verifyExpectationsMet();
-
-		assertThat( failureHandler.genericFailures ).isEmpty();
-
-		List<EntityIndexingFailureContext> entityFailures = failureHandler.entityFailures.get( 2 );
-		awaitFor( () -> assertThat( entityFailures ).hasSize( 1 ) );
-
-		EntityIndexingFailureContext entityFailure = entityFailures.get( 0 );
-		checkId2EntityEventFailure( entityFailure );
-	}
-
-	@Test
-	@TestForIssue(jiraKey = "HSEARCH-4230")
-	public void backendFailure_failedDeleteThenAdd() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-			entity1.setIndexedField( "initialValue" );
-			session.persist( entity1 );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.add( "1", b -> b
-							.field( "indexedField", "initialValue" ) );
-		} );
-		backendMock.verifyExpectationsMet();
-
-		outboxEventFinder.enableFilter( true );
-
-		// Delete the entity (but don't trigger indexing yet: events are being filtered)
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = session.load( IndexedEntity.class, 1 );
-			session.delete( entity1 );
-		} );
-
-		// Remember the events at this point
-		List<Long> eventIdsUpToDelete = new ArrayList<>();
-		withinTransaction( sessionFactory, session -> {
-			eventIdsUpToDelete.addAll( outboxEventFinder.findOutboxEventIdsNoFilter( session ) );
-		} );
-
-		// Re-create the entity (but don't trigger indexing yet: events are being filtered)
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-			entity1.setIndexedField( "updatedValue" );
-			session.persist( entity1 );
-		} );
-
-		// This is the point of this test:
-		// simulate the processing of the delete (which fails) then the second add (which succeeds),
-		// and also the processing of the delete event's retry (which succeeds).
-		// The retry needs to be processed before the add, otherwise the entity will be missing from the index.
-
-		// Delete (failure, schedules a retry)
-		CompletableFuture<?> failingFuture = new CompletableFuture<>();
-		failingFuture.completeExceptionally( new SimulatedFailure( "Delete work on #1 failed!" ) );
-		backendMock.expectWorks( IndexedEntity.INDEX )
-				.createAndExecuteFollowingWorks( failingFuture )
-				.delete( "1" );
-		outboxEventFinder.showOnlyEvents( eventIdsUpToDelete );
-		backendMock.verifyExpectationsMet();
-
-		List<EntityIndexingFailureContext> entityFailures = failureHandler.entityFailures.get( 1 );
-		awaitFor( () -> assertThat( entityFailures ).hasSize( 1 ) );
-
-		// Delete retry + add
-		backendMock.expectWorks( IndexedEntity.INDEX )
-				.addOrUpdate( "1", b -> b
-						.field( "indexedField", "updatedValue" ) );
-		outboxEventFinder.enableFilter( false );
-		backendMock.verifyExpectationsMet();
-	}
-
-	@Test
-	public void backendFailure_twoFailuresOfTheSameIndexingWork() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-			entity1.setIndexedField( "initialValue" );
-			session.persist( entity1 );
-
-			IndexedEntity entity2 = new IndexedEntity();
-			entity2.setId( 2 );
-			entity2.setIndexedField( "initialValue" );
-			session.persist( entity2 );
-
-			IndexedEntity entity3 = new IndexedEntity();
-			entity3.setId( 3 );
-			entity3.setIndexedField( "initialValue" );
-			session.persist( entity3 );
-
-			CompletableFuture<?> failingFuture = new CompletableFuture<>();
-			failingFuture.completeExceptionally( new SimulatedFailure( "Indexing work #2 failed!" ) );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.add( "1", b -> b
-							.field( "indexedField", "initialValue" )
-					)
-					.add( "3", b -> b
-							.field( "indexedField", "initialValue" )
-					)
-					.createAndExecuteFollowingWorks( failingFuture )
-					.add( "2", b -> b
-							.field( "indexedField", "initialValue" )
-					)
-					// retry (fails too):
-					.addOrUpdate( "2", b -> b
-							.field( "indexedField", "initialValue" )
-					);
-
-			// finally it works:
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.addOrUpdate( "2", b -> b
-							.field( "indexedField", "initialValue" )
-					);
-
-		} );
-		backendMock.verifyExpectationsMet();
-
-		assertThat( failureHandler.genericFailures ).isEmpty();
-
-		List<EntityIndexingFailureContext> entityFailures = failureHandler.entityFailures.get( 2 );
-		awaitFor( () -> assertThat( entityFailures ).hasSize( 2 ) );
-
-		EntityIndexingFailureContext entityFailure = entityFailures.get( 0 );
-		checkId2EntityEventFailure( entityFailure );
-
-		entityFailure = entityFailures.get( 1 );
-		checkId2EntityEventFailure( entityFailure );
-	}
-
-	@Test
-	public void backendFailure_numberOfTrialsExhausted() {
-		OrmUtils.withinTransaction( sessionFactory, session -> {
-			IndexedEntity entity1 = new IndexedEntity();
-			entity1.setId( 1 );
-			entity1.setIndexedField( "initialValue" );
-			session.persist( entity1 );
-
-			IndexedEntity entity2 = new IndexedEntity();
-			entity2.setId( 2 );
-			entity2.setIndexedField( "initialValue" );
-			session.persist( entity2 );
-
-			IndexedEntity entity3 = new IndexedEntity();
-			entity3.setId( 3 );
-			entity3.setIndexedField( "initialValue" );
-			session.persist( entity3 );
-
-			CompletableFuture<?> failingFuture = new CompletableFuture<>();
-			failingFuture.completeExceptionally( new SimulatedFailure( "Indexing work #2 failed!" ) );
-
-			backendMock.expectWorks( IndexedEntity.INDEX )
-					.add( "1", b -> b
-							.field( "indexedField", "initialValue" )
-					)
-					.add( "3", b -> b
-							.field( "indexedField", "initialValue" )
-					)
-					.createAndExecuteFollowingWorks( failingFuture )
-					.add( "2", b -> b
-							.field( "indexedField", "initialValue" )
-					)
-					// retry (fails too):
-					.createAndExecuteFollowingWorks( failingFuture )
-					.addOrUpdate( "2", b -> b
-							.field( "indexedField", "initialValue" )
-					)
-					// retry (fails too):
-					.createAndExecuteFollowingWorks( failingFuture )
-					.addOrUpdate( "2", b -> b
-							.field( "indexedField", "initialValue" )
-					);
-
-			// no more retry
-		} );
-		backendMock.verifyExpectationsMet();
-
-		assertThat( failureHandler.genericFailures ).isEmpty();
-
-		List<EntityIndexingFailureContext> entityFailures = failureHandler.entityFailures.get( 2 );
-		awaitFor( () -> assertThat( entityFailures ).hasSize( 4 ) );
-
-		for ( int i = 0; i < 3; i++ ) {
-			EntityIndexingFailureContext entityFailure = entityFailures.get( i );
-			checkId2EntityEventFailure( entityFailure );
-		}
-
-		EntityIndexingFailureContext entityFailure = entityFailures.get( 3 );
-		assertThat( entityFailure.failingOperation() ).isEqualTo( "Processing an outbox event." );
-		assertThat( entityFailure.throwable() )
-				.isInstanceOf( SearchException.class )
-				.hasMessageContaining( "Max '3' retries exhausted to process the event. Event will be lost." );
-		hasOneReference( entityFailure.entityReferences(), "indexed", 2 );
-	}
-
-	private void checkId2EntityEventFailure(EntityIndexingFailureContext entityFailure) {
-		assertThat( entityFailure.failingOperation() ).isEqualTo( "Processing an outbox event." );
-		assertThat( entityFailure.throwable() )
-				.isInstanceOf( SimulatedFailure.class )
-				.hasMessageContaining( "Indexing work #2 failed!" );
-		hasOneReference( entityFailure.entityReferences(), "indexed", 2 );
-	}
-
-	@SuppressWarnings("unchecked")
-	private void hasOneReference(List<Object> entityReferences, String entityName, Object id) {
-		assertThat( entityReferences ).hasSize( 1 );
-		EntityReference entityReference = (EntityReference) entityReferences.get( 0 );
-		assertThat( entityReference.name() ).isEqualTo( entityName );
-		assertThat( entityReference.id() ).isEqualTo( id );
-	}
-
-	private static void awaitFor(ThrowingRunnable assertion) {
-		await().timeout( 2, TimeUnit.SECONDS ).untilAsserted( assertion );
 	}
 
 	@Entity(name = "indexed")
@@ -442,41 +158,6 @@ public class DatabasePollingAutomaticIndexingEdgeCasesIT {
 
 		public void setIndexedField(String indexedField) {
 			this.indexedField = indexedField;
-		}
-	}
-
-	private static class SimulatedFailure extends RuntimeException {
-		SimulatedFailure(String message) {
-			super( message );
-		}
-	}
-
-	private static class TestFailureHandler implements FailureHandler {
-		private LogFailureHandler delegate = new LogFailureHandler();
-		// there are no concurrent write in this test,
-		// using volatile list / concurrent hashmap only to make the changes on value lists visible by the main thread
-		private volatile List<FailureContext> genericFailures = new ArrayList<>();
-		private Map<Integer, List<EntityIndexingFailureContext>> entityFailures = new ConcurrentHashMap<>();
-
-		@Override
-		public void handle(FailureContext context) {
-			genericFailures.add( context );
-			// For easier debugging
-			delegate.handle( context );
-		}
-
-		@Override
-		@SuppressWarnings("unchecked")
-		public void handle(EntityIndexingFailureContext context) {
-			for ( Object item : context.entityReferences() ) {
-				EntityReference entityReference = (EntityReference) item;
-				Integer id = (Integer) entityReference.id();
-
-				entityFailures.computeIfAbsent( id, key -> new ArrayList<>() );
-				entityFailures.get( id ).add( context );
-			}
-			// For easier debugging
-			delegate.handle( context );
 		}
 	}
